+++ src/Collection/AuditedCollection.php
@@ -84,7 +84,7 @@ class AuditedCollection implements Collection
      * @return true
      */
     #[\ReturnTypeWillChange]
-    public function add($element)
+    public function add(mixed $element)
     {
         throw new AuditedCollectionException('The AuditedCollection is read-only');
     }
@@ -100,7 +100,7 @@ class AuditedCollection implements Collection
      * @return bool
      */
     #[\ReturnTypeWillChange]
-    public function contains($element)
+    public function contains(mixed $element)
     {
         $this->forceLoad();
 
@@ -123,6 +123,8 @@ class AuditedCollection implements Collection
     }
 
     /**
+     * @param string|int $key
+     *
      * @return T|null
      */
     #[\ReturnTypeWillChange]
@@ -135,13 +137,17 @@ class AuditedCollection implements Collection
      * @return bool
      */
     #[\ReturnTypeWillChange]
-    public function removeElement($element)
+    public function removeElement(mixed $element)
     {
         throw new AuditedCollectionException('Audited collections does not support removal');
     }
 
     /**
+     * @param int|string $key
+     *
      * @return bool
+     *
+     * @phpstan-param TKey $key
      */
     #[\ReturnTypeWillChange]
     public function containsKey($key)
@@ -152,6 +158,8 @@ class AuditedCollection implements Collection
     }
 
     /**
+     * @param string|int $key
+     *
      * @return object
      *
      * @phpstan-return T
@@ -346,7 +354,7 @@ class AuditedCollection implements Collection
      * @return TKey|false
      */
     #[\ReturnTypeWillChange]
-    public function indexOf($element)
+    public function indexOf(mixed $element)
     {
         $this->forceLoad();
 
@@ -354,6 +362,9 @@ class AuditedCollection implements Collection
     }
 
     /**
+     * @param int      $offset
+     * @param int|null $length
+     *
      * @return object[]
      *
      * @phpstan-return array<TKey,T>
@@ -381,7 +392,7 @@ class AuditedCollection implements Collection
      * @return bool
      */
     #[\ReturnTypeWillChange]
-    public function offsetExists($offset)
+    public function offsetExists(mixed $offset)
     {
         $this->forceLoad();
 
@@ -394,7 +405,7 @@ class AuditedCollection implements Collection
      * @phpstan-return T
      */
     #[\ReturnTypeWillChange]
-    public function offsetGet($offset)
+    public function offsetGet(mixed $offset)
     {
         if ($this->loadedEntities->offsetExists($offset)) {
             $entity = $this->loadedEntities->offsetGet($offset);
@@ -417,12 +428,12 @@ class AuditedCollection implements Collection
         return $resolvedEntity;
     }
 
-    public function offsetSet($offset, $value): void
+    public function offsetSet(mixed $offset, mixed $value): void
     {
         throw new AuditedCollectionException('AuditedCollection is read-only');
     }
 
-    public function offsetUnset($offset): void
+    public function offsetUnset(mixed $offset): void
     {
         throw new AuditedCollectionException('Audited collections does not support removal');
     }