+++ tests/App/AppKernel.php
@@ -22,7 +22,6 @@ use Symfony\Bundle\SecurityBundle\SecurityBundle;
 use Symfony\Bundle\TwigBundle\TwigBundle;
 use Symfony\Component\Config\Loader\LoaderInterface;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
-use Symfony\Component\HttpFoundation\InputBag;
 use Symfony\Component\HttpKernel\Kernel;
 use Symfony\Component\Routing\Loader\Configurator\RoutingConfigurator;
 use Symfony\Component\Security\Http\Attribute\IsGranted;
@@ -58,12 +57,7 @@ final class AppKernel extends Kernel
         return __DIR__;
     }
 
-    /**
-     * TODO: add typehint when support for Symfony < 5.1 is dropped.
-     *
-     * @param RoutingConfigurator $routes
-     */
-    protected function configureRoutes($routes): void
+    protected function configureRoutes(RoutingConfigurator $routes): void
     {
         $routes->import(sprintf('%s/config/routes.yml', $this->getProjectDir()));
     }
@@ -74,18 +68,8 @@ final class AppKernel extends Kernel
 
         $loader->load(__DIR__.'/config/config.yml');
 
-        if (class_exists(IsGranted::class)) {
-            $loader->load(__DIR__.'/config/config_symfony_v6.yml');
-        } elseif (class_exists(InputBag::class)) {
+        if (!class_exists(IsGranted::class)) {
             $loader->load(__DIR__.'/config/config_symfony_v5.yml');
-        } else {
-            $loader->load(__DIR__.'/config/config_symfony_v4.yml');
-        }
-
-        if (version_compare(\PHP_VERSION, '8.0.0', '>=')) {
-            $loader->load(__DIR__.'/config/config_php8.yml');
-        } else {
-            $loader->load(__DIR__.'/config/config_php7.yml');
         }
 
         $loader->load(__DIR__.'/config/services.php');